Nico friend location

+++ findme/ContactViewController.swift
@@ -70,8 +70,8 @@ class ContactViewController: UITableViewController, NSURLConnectionDelegate{
                 for user : NSDictionary in json{
                     let name = user["pseudo"] as? String
                     let password = user["password"] as? String
-                    let x = user["x"] as? Float
-                    let y = user["y"] as? Float
+                    let x = user["x"] as? Double
+                    let y = user["y"] as? Double
                     let friendList = user["friendList"] as? [User]
                     let jsonUser = User(pseudo: name!, x: x!, y: y!, password: password!, friendList: friendList!)
                     self.users.append(jsonUser)
